+++ src/mesh/iter.rs
@@ -86,25 +86,9 @@ impl<'a> Iterator for Primitives<'a> {
 
 fn map_morph_target<'a>(
     document: &'a crate::Document,
-    json: &json::mesh::MorphTarget,
+    json: &'a json::mesh::MorphTarget,
 ) -> MorphTarget<'a> {
-    let positions = json
-        .positions
-        .as_ref()
-        .map(|index| document.accessors().nth(index.value()).unwrap());
-    let normals = json
-        .normals
-        .as_ref()
-        .map(|index| document.accessors().nth(index.value()).unwrap());
-    let tangents = json
-        .tangents
-        .as_ref()
-        .map(|index| document.accessors().nth(index.value()).unwrap());
-    MorphTarget {
-        positions,
-        normals,
-        tangents,
-    }
+    MorphTarget { document, json }
 }
 
 impl<'a> ExactSizeIterator for MorphTargets<'a> {}